refactor(cli): split mastra score event importer
